Non-highlighted cells have a small diameter.
I reorganized the plotData code to calculate what cells should be
highlighted and then to set up all the data point parameters at once.
Unfortunately the performance suffers later in the time series because
x3dom does not recognize the scale as a dynamic variable and all of the
points have to be redrawn.