loads cell position data from the new file format
I changed the code to load the cell position data from the new cell
position/gene expression csv file. It does not currently work to load
the entire time series, as it tries to read in the whole file at once
(which takes up too much memory, at least on my machine), so I got it
working on a subset of the rows (for example the first 100 time points
works well). Next step will be to optimize the data loading to work on a
streaming basis, row by row, or to break the data into chunks. I also
had to do a lot of data wrangling to get the positions set up. I will
send an email around with more details on what I did to the cell
positions, along with the updated data file as an attachment.