Updates to highlight coloring
I refactored the highlighting code so that now the color of the cell is
an attribute of the metadata object in the cellmap data structure. This
helps to speed up the plotting code because the highlights are not
recomputed every time plotData() is called. Instead, the colors of the
cells (and an attribute of each cell's metadata that indicates whether
or not it is selected to be highlighted) are only updated when the user
changes their highlight selections.