user selection of individual points
When a user clicks on either a 3D point or a node in the lineage tree,
the corresponding node in the tree and any corresponding data point in
the 3D plot will be highlighted with a yellow circle. To get this
working, I also refactored some more of the display code so that we no
longer have to re-plot the entire vis for every update. Updates to
lineage highlights for data point size/color now only affect the data
points that must change. I also began storing certain display
parameters, including color and whether or not a cell is
highlighted/selected on the cellmap objects themselves so that the
plotting functions can simply look up the correct color and status of a
point at the time of plotting instead of recalculating these things.