More work on gene expression plot
I tried a whole bunch of ways to optimize the gene expression plot, and
mostly failed. It is perhaps slightly more efficient, and I worked out
some bugs in getting the right data to display at the right time, but
the vis still slows down drastically once there are more than a few
cells to keep track of. I also added ordinal scales to the plot, but the
plot changes size at different time points due to the number of
available pixels not evenly dividing among the dimensions of the plot. I
still have some ideas for improving it, but I don't have any more time
to work on this, so for now I will just put them into the issue on
GitHub for anyone to try.