more gene expression updates
I tried one more idea to make the gene expression data more efficient.
The vis now only stores the full gene expression pattern for the first
time point in which a cell exists. At subsequent time points for that
cell, only the indices of differences from the original gene expression
pattern are stored. This saves space, but does not help all that much in
terms of efficiency of display. I also cleaned up the documentation of
the new code and removed commented out code.