load new csv file with PCA columns
I moved the PCA values to be columns just to the left of all of the
expression values and changed the data loading code to accommodate this
and store the PCA1,2,3 values in the csvdata structure. I also am
including the code I used to transform the binary csv file that Andrew
sent with PCA columns on the end of each line.