updates to gene expression plotting
I tried implementing the heatmap using rows (cells) as the d3 data
hoping that it would render faster since entire rows can be
added/removed per time point. It seems to be a bit faster, but not as
much as I had hoped. One problem is that at each time point the heights
and positions of the rectangles still must be rescaled due to
subdivision of space when other cells divide. It would be nice to scale
the height of all rect objects to the height of their group container,
although it's not clear to me how this works with SVG.