fixes gene expression color and lineage sorting
I fixed a small bug preventing highlights from appearing on the gene
expression heatmap, and I added a comparator function that properly
sorts cell names based on the canonical ordering of the lineage tree.
This comparator works for sorting both the gene expression plot rows and
the branches of the lineage tree plot.