gene expression plot even faster, layout improvements
I tweaked the gene expression plot some more, and it's pretty passable
now in terms of speed. It will animate right through all the times on
the slow speed setting, but I haven't tested on faster settings. While
working on this, I also went back and worked out some more bugs from the
gene expression data file, I'll send the updated data out right after
this commit.

I also worked on getting the layout closer to what it will have to be in
the end. This involved working with some more divs and css, and swapping
the x and y axes on the gene expression plot.